// src/app/interfaces/add-new-match.type.ts
export type AddNewMatch = {
	playerHome: string;
	playerAway: string;
	set1HomeScore: number;
	set1AwayScore: number;
	set2HomeScore: number;
	set2AwayScore: number;
	set3HomeScore: number;
	set3AwayScore: number;
	set4HomeScore?: number;
	set4AwayScore?: number;
	set5HomeScore?: number;
	set5AwayScore?: number;
};